requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
onnxruntime
opencv-python-headless
ffmpeg-python